Training mode added to launcher

+++ Project Aces URP/Assets/Scripts/GamemodeData.cs
@@ -7,10 +7,10 @@ public class GamemodeData : ScriptableObject
     public string gamemodeName = "Name of Gamemode";
     public bool EliminationPoints = false;
     public bool isTeams = true;
-    public int playerCount;
-    public int AIPlayers;
-    public int scoreValue;
-    public int maxScore;
-    public int timeLimit;
+    public byte playerCount;
+    public byte AIPlayers;
+    public byte scoreValue;
+    public byte maxScore;
+    public byte timeLimit;
     
 }
